stopcode bugfix
Kind of a bandaid that anecdotally fixes the bug of the stop not getting added about a tenth of the time. Even with the bug, the data would still INFLATE fine, but infgen would still report the error. This patch fixes that bug, and no further bugs have cropped up, in practice. But this doesn't mean that new bugs might be introduced, as it is a hack